+++ judo-runtime-core-jsl-itest/pom.xml
@@ -421,6 +421,21 @@
                             <destination>${basedir}/target/generated-test-sources/sdk-core/TestStaticQueriesWithDefaultValues</destination>
                         </configuration>
                     </execution>
+
+                    <execution>
+                        <id>execute-psm-test-model-from-artifact-TransferSDKTestModel</id>
+                        <phase>generate-test-sources</phase>
+                        <goals>
+                            <goal>generate</goal>
+                        </goals>
+                        <configuration>
+                            <psm>
+                                ${basedir}/target/generated-test-sources/model/TransferSDKTestModel-psm.model
+                            </psm>
+                            <destination>${basedir}/target/generated-test-sources/sdk-core/TransferSDKTestModel</destination>
+                        </configuration>
+                    </execution>
+
                 </executions>
 
             </plugin>
@@ -462,6 +477,7 @@
                                 <source>${project.basedir}/target/generated-test-sources/sdk-core/TestStaticQueriesWithConstantParameters</source>
                                 <source>${project.basedir}/target/generated-test-sources/sdk-core/OperatorAndFunctionOfEnumTest</source>
                                 <source>${project.basedir}/target/generated-test-sources/sdk-core/TestStaticQueriesWithDefaultValues</source>
+                                <source>${project.basedir}/target/generated-test-sources/sdk-core/TransferSDKTestModel</source>
                             </sources>
                         </configuration>
                    </execution>
